refactor!: unify slash handling
It is hard to remember to add `/` at the end of the URL when configuring
a new etherpad URL. Also, web server proxies tend to do weird stuff when
you assume that you have a slash on the end of each URL.

So, I took a look at how to avoid doing that. It turns out that both
urljoin / os.path.join are kinda bad for handling URL building. Also,
APPLICATION_ROOT didn't seem that necessary since Flask knows what to do
with a `/` at the start of a URL, so I dropped it.

I think this doesn't break anything!